requirements.txt
logging==0.4.9.6
Markdown==3.8.2
python-dotenv==1.1.1
requests==2.32.5
vk_api==11.10.0


